Remove waitForUpdate from tests to make them more deterministic
