+++ news/3 Code Health/11058.md
@@ -1,0 +1,1 @@
+Make functional tests more deterministic
