+++ src/client/datascience/interactive-common/synchronization.ts
@@ -145,6 +145,7 @@ const messageWithMessageTypes: MessageMapping<IInteractiveWindowMapping> & Messa
     [InteractiveWindowMessages.NotebookRunSelectedCell]: MessageType.other,
     [InteractiveWindowMessages.OpenLink]: MessageType.other,
     [InteractiveWindowMessages.OpenSettings]: MessageType.other,
+    [InteractiveWindowMessages.OutputToggled]: MessageType.other,
     [InteractiveWindowMessages.ProvideCompletionItemsRequest]: MessageType.other,
     [InteractiveWindowMessages.ProvideCompletionItemsResponse]: MessageType.other,
     [InteractiveWindowMessages.ProvideHoverRequest]: MessageType.other,
@@ -164,6 +165,7 @@ const messageWithMessageTypes: MessageMapping<IInteractiveWindowMapping> & Messa
     [InteractiveWindowMessages.SaveAll]: MessageType.other,
     [InteractiveWindowMessages.SavePng]: MessageType.other,
     [InteractiveWindowMessages.ScrollToCell]: MessageType.syncWithLiveShare,
+    [InteractiveWindowMessages.SelectedCell]: MessageType.other,
     [InteractiveWindowMessages.SelectJupyterServer]: MessageType.other,
     [InteractiveWindowMessages.SelectKernel]: MessageType.other,
     [InteractiveWindowMessages.SendInfo]: MessageType.other,
