+++ src/datascience-ui/interactive-common/redux/store.ts
@@ -135,6 +135,7 @@ function createTestMiddleware(): Redux.Middleware<{}, IStore> {
     // Make sure all dynamic imports are loaded.
     const transformPromise = forceLoad();
 
+    // tslint:disable-next-line: cyclomatic-complexity
     return (store) => (next) => (action) => {
         const prevState = store.getState();
         const res = next(action);
@@ -154,6 +155,10 @@ function createTestMiddleware(): Redux.Middleware<{}, IStore> {
         if (previousSelection.focusedCellId !== currentSelection.focusedCellId && currentSelection.focusedCellId) {
             // Send async so happens after render state changes (so our enzyme wrapper is up to date)
             sendMessage(InteractiveWindowMessages.FocusedCellEditor, { cellId: action.payload.cellId });
+        }
+        if (previousSelection.selectedCellId !== currentSelection.selectedCellId && currentSelection.selectedCellId) {
+            // Send async so happens after render state changes (so our enzyme wrapper is up to date)
+            sendMessage(InteractiveWindowMessages.SelectedCell, { cellId: action.payload.cellId });
         }
         // Special case for unfocusing a cell
         if (previousSelection.focusedCellId !== currentSelection.focusedCellId && !currentSelection.focusedCellId) {
@@ -202,6 +207,14 @@ function createTestMiddleware(): Redux.Middleware<{}, IStore> {
             const diff = afterFinished.filter((r) => prevFinished.indexOf(r) < 0);
             // Send async so happens after the render is actually finished.
             sendMessage(InteractiveWindowMessages.ExecutionRendered, { ids: diff });
+        }
+
+        // Hiding/displaying output
+        const prevHidingOutput = prevState.main.cellVMs.filter((c) => c.hideOutput);
+        const afterHidingOutput = afterState.main.cellVMs.filter((c) => c.hideOutput);
+        if (!fastDeepEqual(prevHidingOutput, afterHidingOutput)) {
+            // Send async so happens after the render is actually finished.
+            sendMessage(InteractiveWindowMessages.OutputToggled);
         }
 
         // Entering break state in a native cell
