+++ src/test/datascience/plotViewer.functional.test.tsx
@@ -5,7 +5,7 @@
 import '../../client/common/extensions';
 
 import * as assert from 'assert';
-import { mount, ReactWrapper } from 'enzyme';
+import { ComponentClass, mount, ReactWrapper } from 'enzyme';
 import { parse } from 'node-html-parser';
 import * as React from 'react';
 import { Disposable } from 'vscode';
@@ -13,7 +13,6 @@ import { Disposable } from 'vscode';
 import { IPlotViewerProvider } from '../../client/datascience/types';
 import { MainPanel } from '../../datascience-ui/plot/mainPanel';
 import { DataScienceIocContainer } from './dataScienceIocContainer';
-import { waitForUpdate } from './reactHelpers';
 
 // import { asyncDump } from '../common/asyncDump';
 suite('DataScience PlotViewer tests', () => {
@@ -35,6 +34,89 @@ suite('DataScience PlotViewer tests', () => {
         plotViewerProvider = ioc.get<IPlotViewerProvider>(IPlotViewerProvider);
 
         return ioc.wrapper!;
+    }
+
+    function waitForComponentDidUpdate<P, S, C>(component: React.Component<P, S, C>): Promise<void> {
+        return new Promise((resolve, reject) => {
+            if (component) {
+                let originalUpdateFunc = component.componentDidUpdate;
+                if (originalUpdateFunc) {
+                    originalUpdateFunc = originalUpdateFunc.bind(component);
+                }
+
+                // tslint:disable-next-line:no-any
+                component.componentDidUpdate = (prevProps: Readonly<P>, prevState: Readonly<S>, snapshot?: any) => {
+                    // When the component updates, call the original function and resolve our promise
+                    if (originalUpdateFunc) {
+                        originalUpdateFunc(prevProps, prevState, snapshot);
+                    }
+
+                    // Reset our update function
+                    component.componentDidUpdate = originalUpdateFunc;
+
+                    // Finish the promise
+                    resolve();
+                };
+            } else {
+                reject('Cannot find the component for waitForComponentDidUpdate');
+            }
+        });
+    }
+
+    function waitForRender<P, S, C>(component: React.Component<P, S, C>, numberOfRenders: number = 1): Promise<void> {
+        // tslint:disable-next-line:promise-must-complete
+        return new Promise((resolve, reject) => {
+            if (component) {
+                let originalRenderFunc = component.render;
+                if (originalRenderFunc) {
+                    originalRenderFunc = originalRenderFunc.bind(component);
+                }
+                let renderCount = 0;
+                component.render = () => {
+                    let result: React.ReactNode = null;
+
+                    // When the render occurs, call the original function and resolve our promise
+                    if (originalRenderFunc) {
+                        result = originalRenderFunc();
+                    }
+                    renderCount += 1;
+
+                    if (renderCount === numberOfRenders) {
+                        // Reset our render function
+                        component.render = originalRenderFunc;
+                        resolve();
+                    }
+
+                    return result;
+                };
+            } else {
+                reject('Cannot find the component for waitForRender');
+            }
+        });
+    }
+
+    async function waitForUpdate<P, S, C>(
+        wrapper: ReactWrapper<P, S, C>,
+        mainClass: ComponentClass<P>,
+        numberOfRenders: number = 1
+    ): Promise<void> {
+        const mainObj = wrapper.find(mainClass).instance();
+        if (mainObj) {
+            // Hook the render first.
+            const renderPromise = waitForRender(mainObj, numberOfRenders);
+
+            // First wait for the update
+            await waitForComponentDidUpdate(mainObj);
+
+            // Force a render
+            wrapper.update();
+
+            // Wait for the render
+            await renderPromise;
+
+            // Force a render
+            wrapper.update();
+        }
     }
 
     teardown(async () => {
